% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffviewer-package.R
\docType{package}
\name{diffviewer-package}
\alias{diffviewer}
\alias{diffviewer-package}
\title{diffviewer: HTML Widget to Show File Differences}
\description{
A HTML widget that shows differences between files (text, images, and data frames).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://diffviewer.r-lib.org}
  \item \url{https://github.com/r-lib/diffviewer}
  \item Report bugs at \url{https://github.com/r-lib/diffviewer/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co}

Authors:
\itemize{
  \item Joshua Kunst
  \item Winston Chang \email{winston@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Paul Fitzpatrick (Author of included daff.js library) [copyright holder]
  \item Rodrigo Fernandes (Author of included diff2html library) [copyright holder]
  \item JQuery Foundation (Author of included jquery library) [copyright holder]
  \item Kevin Decker (Author of included jsdiff library) [copyright holder]
  \item Matthew Holt (Author of incldued PapaParse library) [copyright holder]
  \item Huddle (Author of included resemble library) [copyright holder]
}

}
\keyword{internal}
