% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-diff.R
\name{diffviewer_widget}
\alias{diffviewer_widget}
\title{Create an htmlwidget that shows differences between files or directories}
\usage{
diffviewer_widget(old, new, width = NULL, height = NULL, pattern = NULL)
}
\arguments{
\item{old, new}{Names of the old and new directories to compare.
Alternatively, they can be a character vectors of specific files to
compare.}

\item{width, height}{Width and height of widget.}

\item{pattern}{A filter to apply to the old and new directories.}
}
\description{
This function can be used for viewing differences between current test
results and the expected results
}
\examples{
path1 <- tempfile()
path2 <- tempfile()

png(path1)
plot(1:10)
dev.off()

png(path2)
plot(1:10, xlab = "x")
dev.off()

diffviewer_widget(path1, path2)
}
